requirements.txt
pydantic>=2.6
langchain
langchain-openai
python-dotenv
